# app/Models/Cars.php
<?php

namespace App\Models;
use App\Models\Pilot;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use OpenApi\Annotations as OA;

/**
 * @OA\Schema(
 *     schema="Cars",
 *     type="object",
 *
 *     @OA\Property(
 *          property="id",
 *          type="integer",
 *          example="1"
 *     ),
 *
 *     @OA\Property(
 *          property="name",
 *          type="string",
 *          example="BMW"
 *     ),
 *
 *     @OA\Property(
 *          property="pilot_id",
 *          type="integer",
 *          example="1"
 *     ),
 *
 *     @OA\Property(
 *          property="created_at",
 *          type="string",
 *          format="date-time",
 *          example="2022-05-13T00:00:00Z"
 *     ),
 *
 *     @OA\Property(
 *          property="updated_at",
 *          type="string",
 *          format="date-time",
 *          example="2022-05-13T00:00:00Z"
 *     )
 * )
 */
class Cars extends Model
{
    use HasFactory;

    protected $fillable = ['name', 'pilot_id'];
    public function cars_of_pilot()
    {
        return $this->belongsTo(Pilot::class);
    }

}